requirements.txt
# --------- pytorch --------- #
torch
torchvision
lightning>=2.4.0
torchmetrics>=1.4.2
torchshow
git+https://github.com/dr-aheydari/SoftAdapt.git
# --------- Computer vision --------- #
opencv-python

# --------- data --------- #
numpy<2.2.0
scipy
seaborn

# --------- hydra --------- #
hydra-core==1.3.2
# hydra-ax-sweeper==1.2.0
ax-platform==0.4.3
# gpytorch==1.9.0
# botorch
hydra-colorlog

# --------- loggers --------- #
wandb
# neptune-client
# mlflow
# comet-ml
# aim>=3.16.2  # no lower than 3.16.2, see https://github.com/aimhubio/aim/issues/2550

# --------- others --------- #
pre-commit      # hooks for applying linters on commit
rich            # beautiful text formatting in terminal
pytest          # tests
# sh            # for running bash commands in some tests (linux/macos only)
